refactor: rename options file
