+++ src/queries/a11yState.ts
@@ -12,7 +12,7 @@ import type {
   QueryAllByQuery,
   QueryByQuery,
 } from './makeQueries';
-import { AccessibilityOption } from './accessibilityOption';
+import { AccessibilityOption } from './options';
 
 const queryAllByA11yState = (
   instance: ReactTestInstance
