+++ src/queries/role.ts
@@ -15,7 +15,7 @@ import type {
   QueryAllByQuery,
   QueryByQuery,
 } from './makeQueries';
-import { AccessibilityOption } from './accessibilityOption';
+import { AccessibilityOption } from './options';
 
 type ByRoleOptions = {
   name?: TextMatch;
