+++ src/queries/text.ts
@@ -18,7 +18,7 @@ import type {
   QueryAllByQuery,
   QueryByQuery,
 } from './makeQueries';
-import { AccessibilityOption } from './accessibilityOption';
+import { AccessibilityOption } from './options';
 
 export type TextMatchOptions = AccessibilityOption & {
   exact?: boolean;
